requirements.txt
pytest==4.4.0
allure-pytest
PyAudio
sounddevice
SpeechRecognition
baidu-aip
scikit_image
uiautomator2
numpy
pytesseract
pytest-sugar   #打印进度
pytest-rerunfailures  #失败重试
pytest-ordering   #执行顺序
